server_end/requirements.txt
Flask==3.0.3
tensorflow==2.15.0
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
